Fixed loss of precision during ReadPressure.

+++ src/devices/Bmx280/BmxBase.cs
@@ -241,7 +241,7 @@ namespace Iot.Device.Bmx280
             long pres = CompensatePressure(t);
 
             //Return the temperature as a float value
-            return (pres) / 256;
+            return (double)pres / 256;
         }
 
         /// <summary>
